[FEATURE] Introduce AfterFileCommandProcessedEvent
A new PSR-14 based event "AfterFileCommandProcessedEvent"
is introduced in favour of the existing
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_extfilefunc.php']['processData']
hook, which is now marked as deprecated, along with
its interface "ExtendedFileUtilityProcessDataHookInterface".

The new event features the same data, however
the way they can be accessed is improved.

The previous `$action` and `$cmdArr` arguments
are now available within the `getCommand()` method,
while the action is the array key and the command
data ($cmdArr) is the value. This is more in line
with the actual request.

The hooks `$result` argument always contained all
results of previous operations, making it unnecessary
complicated to find out the actual result for the
current operation. This is now resolved. The new
`getResult()` method only returns the result for
the currently executed operation.

Finally the "getConflictMode()" method just returns
the conflict mode, used for the current operation.

Resolves: #95089
Releases: master
Change-Id: I9991eb73aed873da5987f88d2d0255764274e143
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70874
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>