[BUGFIX] Update feature toggles state at once
When updating feature toggles, the corresponding method
iterates over all feature toggles and checks if those
state changed. If so, previously each changed toggle
state was saved individually, leading to problems
with file caches.

This is now fixed by storing the toggles to be changed
in an intermediate array which is then executed a single
time, only modifying the LocalConfiguration file once.

Additonally, the flash message is improved.

Resolves: #95111
Resolves: #95112
Releases: master
Change-Id: Ieea82d6af5837dd3bbd56abe182b5e7affd9eaa6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70896
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>