[BUGFIX] Don't fail language:update command when translation missing
Currently language:update command fails when
a translation of a extension is missing.

This has been fixed by returning a successful
status code even though the file was not found
on the translation server. With the additional
option '--fail-on-warnings' the command fails
if the download of a language pack was not
successful.

Missing files are still reported in CLI stdout
and it still fails if the server is not reachable.
Also missing translations are now logged.

Resolves: #95700
Resolves: #96054
Resolves: #96055
Resolves: #96053
Releases: master, 11.5
Change-Id: I9a19de919e99b591dcf02a5cf70465ad9c5c76b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72274
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>