[BUGFIX] Load all backend routes in installer controller
The routes added via ExtensionManagementUtility::addModule() are missing
when the UriBuilder/Router is used prior to ext_localconf loading.

Also for this particular case, multiple instances of the Router class
were created, as UriBuilder is loaded from the failsafe installtool
container instead of using an instance from the symfony container
(as created by LateBootService n loadExtLocalconfDatabaseAndExtTables).
This resulted in multiple routers being loaded with different sets of
routes.

This patch ensures that the routes are loaded from a UriBuilder/Router
instance that is fully populated with information from ext_localconf.
Therefore GeneralUtility::makeInstance() is avoided and
UriBuilder is fetched from the late booted container instead.

Resolves: #95806
Releases: main, 11.5
Change-Id: Id57bf09fb7798eac26dbdbc6ed89228c54fbafb9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72463
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>