[TASK] Avoid StandaloneView() in functional tests
StandaloneView() initalizes an ExtbaseRequest in
__construct() and adds it to the rendering context.

This is unfortunate for ViewHelper tests, since it
hides the information if a specific ViewHelper has
a dependency to a request.

This patch replaces all StandaloneView usages in
functional tests with the TemplateView class of
typo3fluid, which doesn't do anything with request.

When needed, request objects are created and added within
the test manually, making this dependency transparent.

The patch adds some minor refactorings of some tests
along the way and fixes a couple of smaller bugs.

Resolves: #96573
Releases: main
Change-Id: If1c0fd9c96db0212ad251391b4b12ee967fdf21b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73043
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>