[FEATURE] Introduce TCA type "password"
In the process of using dedicated TCA types,
the new TCA type "password" is introduced and
replaces "eval=password" and "eval=saltedpassword"
of TCA type "input".

Notable changes to the previous behaviour:
- "trim" is always done in DataHandler and FormEngine
- Password field is rendered as input type=password
- The input field has by default "autocomplete=off"

The TCA type "password" introduces the new configuration
"hashed", which can be set to "false", if the field value
should be saved as plaintext to the database. This
configuration has no effect for the tables "fe_users" and
"be_users".

Resolves: #97104
Releases: main
Change-Id: Ia48296291a61df6802ef21105b38b4f508b7a11c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73832
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>