[TASK] Move TypoLinkCodecService to EXT:core
EXT:core unfortunately has a couple of dependencies
to EXT:frontend. One of them are TypoLinkCodecService.

Since this service does not have any context, only related
to EXT:frontend, it is now moved to EXT:core, resolving this
cross dependency.

This also allows to use the service in the DataHandler
instead of duplicating parts of the code, which was
previously done to prevent any EXT:frontend dependency
in the DataHandler.

Resolves: #97217
Releases: main
Change-Id: Ib24bfb530fd0ac486ac30b09f43717c95c98c292
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73993
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>