[TASK] Use bootstrap badges instead of labels
The "label" class was removed in bootstrap 5 and replaced through
"badge". The "label" class was kept in place for a while to keep
the compatibility as high as possible, but it's time to move on.

Since bootstrap 5 also dropped support for severities on certain
components, we are adding them again for the badge to keep the
convenience as high as possible and avoid promoting the "utility css"
approach that was introduced with these decisions as we try to keep
markup changes as low as possible.

This patch also reverts the SCSS includes back to relative paths.

While it is possible to define additional lookup paths for the parser,
so you can have shorter includes - it comes at a cost. The included
sources are opaque to the user and the IDE, this means there is
no support for auto-completion, quick navigation through included files
and can lead to unintended overrides by the user.

Resolves: #97291
Releases: main
Change-Id: I9007af705a90db567353379146df1d76b2366a62
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74163
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>