[!!!][TASK] Resolve most #[ReturnTypeWillChange] attributes
Sets return types for most classes that implement
PHP SPL classes. Only "mixed" return types are not
set: That would be hard breaking for extensions
that support both core v11 and v12, since it is not
PHP 7.4 compatible. Those places are marked with
an @todo to switch to mixed in v13.

Resolves: #98035
Releases: main
Change-Id: Icd62a69270a6718c8356108e55a0729779609c53
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75310
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>