[!!!][TASK] Move LocalConfiguration to system/settings.php
When running in composer-mode TYPO3 can benefit
from having most PHP code outside of the document root,
making TYPO3 behave more like a modern PHP application.

In addition, the name "LocalConfiguration" sounds very confusing,
and has been in its place for 10 years (TYPO3 v6.0).

The file names are changed, as "Local" is very confusing
for newcomers, as it is actually a "instance-specific" or
"installation-specific" regular configuration / settings file,
paving the way to making it easier to understand TYPO3 in general.

For this reason, when TYPO3 is running in composer mode
with a separate public/ folder, the previous locations

* public/typo3conf/LocalConfiguration.php
* public/typo3conf/AdditionalConfiguration.php

are now located at
* config/system/settings.php
* config/system/additional.php

In non-composer mode the files are now moved to
* typo3conf/system/settings.php
* typo3conf/system/additional.php

Combined with the typo3/cms-composer-installers v5,
the actual typo3conf folder does not need to be created anymore
in Composer Mode for v12+.

Configuration files of TYPO3 projects from previous versions
are automatically moved to the proper file location.

Resolves: #98319
Releases: main
Change-Id: I74b1421c0424ebf12e01a3ecdc9a0dbb9387793e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75858
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>