[BUGFIX] Clear 'setup' in 'Create empty start page' sys_template
The installer has a step to 'Create empty start page'
which adds a default sys_template to a first page.

This sys_template should set 'clear' flag to 3 to clear
both 'constants' and 'setup', which triggers inclusion of
global TypoScript. This is the same as done in BE Template
Info module with "Create template for a new site".

While we are at it, we're streamlining the class code
a bit by removing obsolete annotations, adding some
type hints, using constructor property promotion,
and declaring this internal class final.

Resolves: #98552
Releases: main
Change-Id: Ia17e435b6bd02ff9a9c56c159e43997ebd217393
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76082
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>