[TASK] Avoid testing 'non-writable' directory
Testing 'non-writable' directories is flaky in
real life since it depends on the user that
executes these tests. On gitlab CI, it is useless
a-priori, since gitlab runs as root in containers.
Establishing a vfs for this does not help us much,
especially when it just tests if some exception is
thrown, which is done obviously when reading code.
This vfsStream usage is simply not worth the dependency.

Resolves: #98932
Related: #97762
Releases: main
Change-Id: Ifd813d47607e3760e2e617b93a56a59773bc4e62
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76296
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>