[BUGFIX] Disable language (de)activation with write-protected configuration
The language management module writes into system/settings.php to
either enable or disable a language. With the changes made with #98957,
the language management should also inform about a write-protected
configuration file and remove any options that conflict with this
state.

This patch removes any option to either enable or disable a language and
moves the Fluid partial to a more generic location aimed for multiple
usages across several modules.

Resolves: #99008
Related: #98957
Releases: main
Change-Id: Iae8c75e0727011289c421a66266989f44be76932
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76450
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susi Moog <look@susi.dev>
Reviewed-by: Susi Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>