[FEATURE] Move site settings into settings.yaml
Site Settings were introduced with TYPO3 v10.

In contrast to the rest of the site configuration,
site settings are sane defaults for TypoScript constants,
and consist of simple key value logic.

In order to separate settings, make them
accessible and editable in the TYPO3 Backend,
and distinguish between the functionality between
required site configuration and optional settings,
the "settings" part of the settings are moved to
a separate "settings.yaml" file in the site configuration
folder, which will later-on be editable by the TYPO3 Backend
with the new Settings API introduced in a separate patch.

Resolves: #99047
Releases: main
Change-Id: Idbcb33ccf1c0d8a728b80ad4c242b56ff2d69afd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76491
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>