[TASK] Streamline "versionNumberedFilename"
TYPO3 has two global configuration settings:

$TYPO3_CONF_VARS['BE']['versionNumberInFilename']
which acts as a boolean.

$TYPO3_CONF_VARS['FE']['versionNumberInFilename']
which has three options: '', 'embed' and 'querystring'.

This change streamlines both flags to only allow
boolean:
false => append the filemtime to the asset file
true => change the filename to "myfile.[timestamp].extension"

This means, that FE now has the previous
default option 'querystring' migrated to false,
where as 'embed' is automatically migrated to true.

This is done in an automated migration
via the SilentConfigurationUpgrade service.

Resolves: #99810
Releases: main
Change-Id: Ib4780adfae47d6c6cf18b1a98d239b78de5ffbcb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77673
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>