[BUGFIX] Replace internal removed hook usage with PSR-14 event
With #97451 some hooks has been removed, and replaced with
new PSR-14 events. In the commit message and rst-file one
removed hook has been written wrong, which may be the reason
why it has been undetected that a `ext:redirects` has used
that removed hook.

This change implements the now the PSR-14 event to complete
the transition and restore the auto-chain information modal
after page slug changes.

The removed hook name in the corresponding breaking rst-file
and the extension scanner matcher definition are changed to
the correct name of the removed hook.

Resolves: #99835
Related: #97451
Releases: main
Change-Id: I537397d3849af099da4e32bb1782407aa0ee1be0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77706
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>