[TASK] Drop help-block and optimize spacing in form engine
The spacing for formengine elements have been improved.

The legacy .help-block was replaced with .form-text in
bootstrap 4.6, we are now finally adapting to this change
and replacing all leftover ocurrances of .help-block.

The legacy .control-label that was removed in bootstrap 4.6
with .form-label that will make a reappearance in bootstrap 5.3.

Resolves: #100526
Releases: main
Change-Id: Ibae67dc2f712eca2d5d9e437435ccf67687d7008
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78536
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>