[TASK] Drop unused and unnessesary text-* css classes
The class .text-normal is not used at all and can be dropped.

The class .text-pre-wrap was only used on .alert-message
since it is make sense here to unify the behaviour the applied
styles were moved to .alert-message directly.

The class .text-wrap-allowed provided no benefits for
the context it was rendered in and can be dropped.

Resolves: #100591
Releases: main
Change-Id: Ibf6562265da3010d24b5096a00eb1611d6f77279
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78634
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>