[BUGFIX] Fix undo operation for RowUpdater UpgradeWizards
With the refactoring to attributes, RowUpdaters failed
to be fetched from the UpgradeWizardService, as they present
upgrade wizards in the registry but are not registered via
the UgradeWizard service locator.

To fix this gracefully, we avoid using the service locator for row
updaters, falling back to the identifier.

A next step (feature) would be the introduction of a dedicated
attribute and service locator for RowUpdaters.

Resolves: #100539
Related: #99586
Releases: main, 12.4
Change-Id: I2be4400b51f1ecef8059c3c22cf5da01090883dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79652
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>