[BUGFIX] Add better read-only settings.php handling to SilentUpgrade
The SilentUpgrade handles changes "under the hood" by migrating
old to new settings. If the settings.php is not writable, these
changes fail with an exception. As we cannot predict which side-
effects non-execution may have, we cannot allow further usage of
the install tool before those changes are done. We are now giving
the user a better error message and the option to run silent
upgrade with the normal CLI upgrade command.

Resolves: #101193
Releases: main
Change-Id: Id1f201440a9feb0c5e33374cc781893a352a523d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79593
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>