[BUGFIX] Repair installtool core updater button callback
Button data attributes have been removed from th modal API
and are therefore not automatically assigned.
Instead of an initial and explicit data-action attribute,
we rather imply that the initial intention of the
"Check for core updates" is to check for updates,
and therefore use data-action only for subsequent actions.

Resolves: #100865
Releases: main, 12.4
Change-Id: Iac6138619012dd7f4a340fb4f773b7bbaf543583
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79802
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>