[BUGFIX] Do not consider done `UpgradeWizard` as not found
Existing upgrade wizards marked as done should not be considered as
not existing and therefore not returning a error exit code in the
cli command. This was the case before TYPO3 v12.2, changing that
behavior with #99586 to support register upgrade wizard based on
a PHP attribute.

Postulating that this was a unintended side-effect for single upgrade
execution, this change adds a early check if the upgrade wizards is
marked as done to keep a non-error exit code and outputting a use-full
note to the console.

Note: Executing all possible upgrade-wizards are not related to this
      issue, and therefore do not need any adjustment. But this is
      the reason for keeping the `isWizardDone()` check in method
      `getWizard()`.

This can be tested with:

> bin/typo3 upgrade:run migrateSiteSettings ; echo "$?"

for TYPO3 main or

> bin/typo3 upgrade:run svgFilesSanitization ; echo "$?"

for TYPO3 12.4.x (and 11.5.x as verification)

Resolves: #101354
Related: #99586
Releases: main, 12.4
Change-Id: I126fccb014e418dc2e4f00aa4775753d17e11749
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80027
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>