[TASK] Remove jQuery from Install Tool modules
This patch removes the jQuery usages from all Install Tool modules.
It's done in one go because of the dependency of the modules on a
abstract class that provides methods returning jQuery objects.

The miration of the installer (the module that performs
the initial setup) is left for another patch, as it is
basically a distinct tool and out of scope for this patch.
Also the jquery-based chosen-js library will be replaced separately.

Resolves: #101655
Releases: main
Change-Id: I6c7dc7e156bdd65fbf36d0cab6654e5efb06e7a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80516
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>