[TASK] Use DI in SchemaMigrator
Core SchemaMigrator is the main entry point
of the "Database analyzer" and other DB
schema comparison and update functionality.

The class is used in low level scenarios
where the full container is not available.

The patch aims to have dependencies of
SchemaMigrator injected into the class, as
well as injecting SchemaMigrator into
consuming services. SchemaMigrator is used
in low level scenarios that have only the
failsafe container, we thus need to fiddle
with Services.php to configure injection.

The SchemaMigrator functional test is
refactored along the way, testing-framework
got a pre-patch and needs to be raised.

> composer u typo3/testing-framework

Resolves: #101676
Related: #101669
Releases: main
Change-Id: I3dbc7d7b4a30fc7dbefb3c19cd671e3e4281b906
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80530
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>