[BUGFIX] Make WebserverType an optional argument
There is no downside in making the argument optional.
Although the API is marked as internal, it is in use
in third party code, which would break within a bugfix release.

Logically the API makes total sense with the argument
being optional. If the webserver type is not known or not
provided, no additional file is created.

Resolves: #101868
Releases: main, 12.4
Change-Id: Ie559b8685fc582681befe09ff769724f13639862
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80914
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>