[BUGFIX] Adjust config file path in Install Tool password hint
The file `settings.php`, which stores the value of
$GLOBALS['TYPO3_CONF_VARS']['BE']['installToolPassword'] is
located under `config/system/` in Composer-mode. If a project
is not set up with composer it is `typo3conf/system/`.

The hint next to the Install Tool's login mask spells the wrong
path, which is corrected with this patch. Also the difference
between legacy and composer setups is taken into account, without
exposing information about the projects file structure.

Resolves: #102166
Releases: main, 12.4
Change-Id: I5b2313516c1af9623a9c8de8f1dca7bb46fb8ce0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81425
Reviewed-by: Jan Greth <jan@greth.me>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Jan Greth <jan@greth.me>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Nikita Hovratov <nikita.h@live.de>