[BUGFIX] Abort CLI installation when no databases are available.
Note: This patch doesn't make the command be on feature-parity
with the web-based installer. CLI bails out when the DB doesn't exist;
web is able to create non-existant DB at this stage. This would need
implementation as a follow-up feature.

This patch only addresses not raising an exception on CLI installation.

Resolves: #102322
Releases: main, 12.4
Change-Id: Ic71c71669b010d4fc8e9cb9f229cbdb191dbbee8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81779
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>