[BUGFIX] Create same configuration in web and cli install process
The code for the web installer and the cli installer
is only partly aligned and shared, thus may easily
diverge, leading to different installation results.

The code to create system settings is now put
into SetupService and is used for both installation
methods, so that the cli install now also uses
FactoryConfiguration.php.

Additionally, the following cleanups are performed
to prepare further cleanup of the installation process:

* The cli setup does not set insecure trustedHostPattern any more.
* Silent config and template file changes are removed from setup,
  as conceptually these do nothing on newly created config and
  template files.
* New exceptions are created for config file generation instead
  of misusing a FAL-related exception.
* Absolute paths are made relative to project path before being
  passed to exception.

Resolves: #102403
Releases: main, 12.4
Change-Id: I069e9e9f4136bba56d630eaf09f676b6c5ea7904
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85866
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>