[TASK] Revise wording in TCA ext_tables.php check
The TCA ext_tables.php check in install tool refers to a conceptual
change in TYPO3 6.2 and explains, why TCA overrides should be used
in favor to TCA modifications in `ext_tables.php`.

This change removes the explanation about the conceptual change in
TYPO3 6.2 and adds a brief overview of required changes. Additionally,
all file- and folder-names are now surrounded with the HTML code
tag to add proper visual code highlighting.

Resolves: #102477
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I750b5a7c0f56c5b3b4bb009cc8f5aaed14dcf3a2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81888
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>